alterando

+++ src/main/java/controle/finaceiro/api/domain/transacao/TransacaoDTO.java
@@ -2,12 +2,12 @@ package controle.finaceiro.api.domain.transacao;
 
 import lombok.Data;
 
-import java.time.LocalDate;
+import java.math.BigDecimal;
 
 @Data
 public class TransacaoDTO {
     private Boolean despesa;
-    private Double valor;
+    private BigDecimal valor;
     private String descricao;
     private Boolean pago;
     private String data;
